Moved Telerik ComboBox to a sub-namespace of 'CUITe.Controls.HtmlControls' since it is a HTML control.
