+++ src/CUITe/CUITe.csproj
@@ -175,7 +175,7 @@
     <Compile Include="Controls\HtmlControls\HtmlXml.cs" />
     <Compile Include="Controls\HtmlControls\IHasInnerText.cs" />
     <Compile Include="Controls\ModifierKeysLifetime.cs" />
-    <Compile Include="Controls\TelerikControls\ComboBox.cs" />
+    <Compile Include="Controls\HtmlControls\Telerik\ComboBox.cs" />
     <Compile Include="Controls\UIWindowsSecurityWindow.cs" />
     <Compile Include="Controls\WinControls\WinButton.cs" />
     <Compile Include="Controls\WinControls\WinCalendar.cs" />
